chore: refactor generate endpoint to be post link
